4.21 done

+++ part4/blog/controllers/blogs.js
@@ -2,7 +2,8 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-const logger = require('../utils/logger')
+const config = require('../utils/config')
+//const logger = require('../utils/logger')
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -11,9 +12,18 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog.toJSON())
+  } else {
+    response.status(404).end()
+  }
+})
+
 blogsRouter.post('/', async (request, response,next) => {
   //logger.info('Received token:', request.token)
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, config.SECRET)
   //logger.info('Decoded token:', decodedToken)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
@@ -37,8 +47,19 @@ blogsRouter.post('/', async (request, response,next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === user.id.toString() ){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    response.status(401).json({ error: 'missing premission to do this' })
+  }
 })
